requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# OpenAI Agents SDK (for openai_impl)
openai-agents>=0.9.0

# CrewAI (for crew_impl)
crewai>=0.80.0
crewai-tools>=0.14.0

# LangGraph (for langgraph_impl)
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.6.0
langgraph-checkpoint>=2.0.0

# Browser automation
playwright>=1.40.0

# MCP support
mcp>=0.9.0
